requirements.txt
datasets==2.19.0
openai>=1.40.0
tqdm>=4.66.0
python-dotenv>=1.0.1
